Added Lerna.
Refs #1.

Lerna provides two key features for this project.
1. Allows packages to be linked together and dependent on each other directly in the repo.
2. Executes scripts in all packages in dependency order, allowing a b uild script to run over all packages in the right order.

These two features are necessary for organizing the repo and making packages dependent on each other in a usable and testable manner.

Release functionality might come in handy too, but that's not why I'm adding it for now.

I generated this commit by running:

```shell
npx lerna init
```

Only additional change was adding `lerna` to the NPM scripts so developers don't need to install it globally or use `npx` each time.