Adds testing infrastructure with Jasmine.
Refs #1.

Created the Jasmine config with:

```shell
(cd packages/bxpb-runtime/ && npx jasmine init)
```

Since we're using TypeScript and I wanted to leverage the existing build configuration (or else risk divergence between the build and test commands), I use a test-only `tsconfig.spec.json` which extends the base `tsconfig.json`. This writes to a different output directory so tests can be included without polluting the release `dist/` directory. Jasmine then runs on these files to execute the tests.

I don't see an easy way of debugging tests just yet. I tried setting up VSCode's debugger, but I couldn't find a configuration I was happy with, as it needs to pass the debugging port through `npm run` which executes two subcommands. This could be avoided by running the tests directly and not implicitly building with `npm test`, but I'm not comfortable with that as it is far too easy to forget to build. I could split it into multiple tasks in VSCode's configuration, but then I need an `npm run build-for-test` and `npm run exec-tests` which are just both called by `npm test` which is confusing and awkward and has non-obvious reasons for existing. At the end of all that it binds debugging support to VSCode directly, which I'm not a fan of.

In the end, I could probably get debugging to work as is, but I'm not able to find a developer experience I'm satisfied with. I'll probably set up Karma soon to get a quick edit/refresh cycle with DevTools.