Adds descriptor types a simple strong-typed `serve()` function.
Refs #1.

This defines descriptor types for services and methods as well as a `serve()` function which is able to infer the implementation methods required and type check them accordingly.

The test required a mock service, however there is no easy way of having a `Message` implementation without compiling a real *.proto file, which would complicate Karma significantly or yield wierd rebuilds. I'm avoiding that for now, but it may be necessary to add in the future. This might be possible by having a private `bxpb-testdata` package which builds protobufs in its `npm run build`. Then Lerna would be able to build it as a dependency without requiring changes the Karma config, however Karma's live reload would not automatically rebuild it.

Also updated the test commands in the README to include `--stream`, which streams test output to the terminal.