Convert tests to run with Karma.
Refs #1.

Previously we used Jasmine, however this runs within Node. Using Karma for debug tests and Jasmine for normal tests, means that one set runs in a browser and the other set runs in Node, which creates some discrepancies. By using Karma, all tests are consistently running in a browser. Running in a browsers is preferable to Node, because the browser is closer to the real runtime where bxpb will be used.