Adds `FakeEvent` and updates existing tests to use it.
Refs #1.

I realized there's no need to mock `chrome.runtime.onMessage()` because it is provided in the input. Instead I made a `FakeEvent` class which maintains the listeners and exposes a new `trigger()` function for use in testing.

I also included tests for `FakeEvent`, though I have not found a good way to test that `fail()` is used correctly. I found a Stack Overflow question which does this, but I was not able to get it to work as `jasmine.Env.prototype.execute()` does not appear to execute tests synchronously. I wasn't able to find a good way to wait for all tests to be completed. Instead I left those untested for now, hopefully I can find a better way to cover those cases in the future.