Implements `serve()`.
Refs #1.

This defines the wire format of an RPC request/response pair. It listens to the provided transport to `serve()` and awaits a message. When a message is received, it is validated against the expected wire format (`ProtoRequest`) and the method implementation is identified and called. Most of this is attempting to handle all the possible errors which can be encountered, as all the data provided cannot be trusted to be consistent with the specification.

Proto messages are encoded in base64 because that is the simplest option for now. It probably does not matter for the time being, and proper performance analysis would identify the most ideal message format in the future.

I also changed the method implementation to use the exact method name provided in the `*.proto` file. Most of the time, this will be in pascal case per protobuf conventions, which might be awkward for some JavaScript linters/programmers. For example:

```proto
service Greeter {
  // Name "Greet" is pascal case, normal for *.proto files.
  rpc Greet(GreetRequest) returns (GreetResponse) { }
}
```

```typescript
serve(chrome.runtime.onMessage, GREETER_SERVICE, {
  // Name "Greet" is also pascal case, weird for JavaScript/TypeScript files.
  async Greet(req: GreetRequest): Promise<GreetResponse> {
    return new GreetResponse();
  },
});
```

I think this is simpler as it is does not include an awkward conversion of the RPC name in the `.proto` file, which can be tricky to do correctly. Messages like `JSObject` would naively be turned into `jSObject` which doesn't make much sense. There are lots of edge cases which cannot be handled algorithmically. This is also simpler for developers as there is a direct one-to-one correlation. Downside is that this looks less JavaScript-y and some linters may complain about it. The linter warnings may be annoying, so we'll have to see exactly how well they play with this format.