Adds client runtime.
Refs #1.

This introduces a `ProtoClient` class with an `rpc` function.

Generated clients will extend `ProtoClient` with their implementation. This gives a consistent constructor interface for usage of the clients and allows protected members to be added in the future if necessary.

`rpc()` does the real work of sending a message to the backend service and awaiting the response. Again, most of the work here is validating data, but this turned out to be simpler than the service side.

Both of these APIs are marked via comment as being private. They need to be exported so generated code can call them, but should **not** be called directly by user code. This allows the generated code structure to be entirely an implementation detail which I can change at will. That means the runtime can have a new function added and the compiler can generate code which uses it without any backwards-compatibility concerns. Downside of this decision is that it does not allow client/service code to be independently rolled out. I think this is ok, as browser extensions are released monolithically through a store, so decoupled releases is a pretty specific use case. If this decision becomes a real problem for a significant number of users, then maybe we can revisit this in the future.

`rpc()` could have been a protected method of `ProtoClient`, but I opted for a separate function so it would be harder to misuse, as developers would need to import it before calling it. As long as I expose this code at some kind of `import { rpc } from 'internal/DO_NOT_DEPEND_OR_ELSE/client';`, then I think we should be ok.