Adds background script.
Refs #1.

This adds a background script to the extension manifest so it is loaded on startup.

The build strives to be as simple as possible, so it just uses `tsc` and no bundler. This is accomplished by leveraging native ESM imports from the browser. https://medium.com/front-end-weekly/es6-modules-in-chrome-extensions-an-introduction-313b3fce955b

Typically, background scripts are referenced as JavaScript files from the manifest. In this case, we reference an HTML file, with a `<script type="module">` in order to use ESM imports. Because there is no bundler, one weird aspect is that imports need to include `.js` at the end. Otherwise module resolution in the browser will fail to find the file. See https://github.com/microsoft/TypeScript/issues/16577 and https://github.com/alshdavid/tsc-website/pull/1

I used a subdirectory for TypeScript/JavaScript code rather than including them directly in `src/` and `dist/` because the build process does not output a single file. This meant I had to pick a name for the subdirectory and the best I came up with was `js/`. I wanted the name under `src/` and `dist/` to be the same, so it is a little strange that TypeScript source files live under the `js/` directory, but I've definitely seen weirder naming conventions.

The `tsconfig.json` file came from `(cd examples/greeter/ && npx typescript --init)` with some settings pulled from `packages/bxpb-runtime/` for consistency. This uses inline sourcemaps for simplicity since they do not need to be exported to downstream projects like the runtime.