Adds proto compilation and Rollup build.
Refs #1.

Switched the build pipeline to use Rollup to bundle TypeScript. This is necessary because Protobuf generated code can only use Closure or ComonJS, there is currently no ESM option. As a result I opted for CommonJS and the compilation must work with that format.

This example was intended to be the simplest, using only NPM scripts with no Webpack/Rollup/Grunt/Gulp/etc. However, `tsc` doesn't provide any means of actually using CommonJS without a bundler. As a result, I added Rollup to this example as I felt it was the "simplest", based on my totally subjective personal experience. I've had trouble with Webpack in the past for multiple output JavaScript files with disjoint dependencies, which is exactly the use case I'm trying to support with this library. I'm still using NPM scripts as the main driver for the build, only using Rollup for the TypeScript compilation and bundling.

After all this, I'm able to generate protobuf code and import it into the build. This uses `grpc-tools` which ships a version of `protoc` which is installable via NPM. Otherwise devs have to install `protoc` themselves, which I would much rather avoid. This generates the JavaScript files, but no TypeScript definitions unfortunately. I used `protoc-gen-ts` to generate typings, though it is really sad that I need a community tool just to use protobufs in TypeScript.

One side effect of using the gRPC tool is that it generates a `myprotofile_grpc_pb.d.ts` file, which contains definitions for gRPC and links to that runtime. This is useless for the purposes of this project but I could not find a way to prevent the file from being generated. Instead, I simply delete that file after generation, which is the best I can do.

`protoc` is more awkward to use than I had hoped, as it requires all input files to be explicitly named. I wanted to just say "Compile all protos under the proto/ directory", however there is no way to do this. Instead, I had to use a `find` command to list out all such files, though this does create more of a build dependency on Linux. None of this would work for Windows, though that isn't too much of a concern right now.

I also wanted generated protos to be available under `import * as foo_pb from 'protos/...'`. This requires a `paths` argument in the `tsconfig.json` file, which works fine with `tsc` but not with Rollup. There is probably a way to configure this to be compatible with Rollup, but I couldn't find any obvious documented solution in `@rollup/plugin-typescript`. Instead, attempting to be as simple as possible, I decided to just import from `../`. I think that is a bit of a code smell, but it's the most straightforward and obvious solution for any users who look at this example.

I also converted `background.js` to a simple script as it no longer needs to be a module.