Adds popup build.
Refs #1.

This copies the background Rollup config to apply to the popup as well with a separate top-level file for it.

Code could pretty easily be shared for the Rollup config between the two binaries, but I decided simplicity of the configuration is more important than maintainability in this particular case.