Exports types from `bxpb-runtime`.
Refs #1.

This allows downstream consumers to see relevant Chrome and Protobuf types for exported symbols of `bxpb-runtime`. According to TypeScript docs, such types should be listed as "dependencies" rather than "devDependencies".

https://www.typescriptlang.org/docs/handbook/declaration-files/publishing.html

Note that this doesn't totally seem to work through Lerna, attempting to add `bxpb-runtime` to the example still does not install `@types/chrome` or `@types/google-protobuf`. I think this is an issue with Lerna however, and this is the right way of including such type dependencies for a library project like this. Fortunately the example app needs a direct dependency on both of those types anyways, so it doesn't really matter in this particular case.