Adds Greeter client and service to example.
Refs #1.

This creates a trivial service to implement the Greeter in the background script and adds a client which calls it from the popup.

I had to hand-write the client and service stubs which should be auto-generated by the protobuf compiler. This is tempoary until the compiler plugin is written to do it for me.