Adds `bxpb-protoc-plugin` package.
Refs #1.

This creates a new package for the `protoc` plugin. Also included some documentation in the README, but kept things somewhat vague around the actual output content for now.

I did realize that the compiler should probably output `.d.ts` and `.js` files, or else user code would be responsible for compiling the `.ts`. This would be annoying for JS users and could have conflicting options for TS users.

I also added a README at the root of all packages so they could be listed with short descriptions for each.