Created "Hello World" binary for `protoc` plugin.
Refs #1.

This sets up a simple TypeScript compilation for the `protoc` plugin and exports it from the package.

The binary is simply a symlink to the root JavaScript file with a Shebang for the Node binary. TypeScript can't easily bundle CommonJS for Node, so instead I had `tsc` output to a `dist/` directory, so all the individual files would be preserved. Imports still seem to work as expected.

I included inline sourcemaps for simplicitly, though a user probably would not want that in their compiler invocation, as it is a bunch of extra information. If someone complains, we might want to revisit and use adjacent sourcemap files.