Add Jasmine tests to `bxpb-protoc-plugin`.
Refs #1.

This adds Jasmine to run tests for `bxpb-protoc-plugin`. There is an initial dummy test for the package.

`bxpb-protoc-plugin.ts` as the top-level file is not easy to test, so instead it simply imports from another file and calls a function there which implements all the "real" functionality. That allows all other files to be tested without weird side-effects, while only leaving a few lines in the top-level file untested. The file is intended to be so trivial as to not require tests.

I used the multiple `tsconfig.json` files trick again for compiling the binary itself and the tests. This allows the binary to not include test code, while the tests include everything.