Documents how to debug `bxpb-protoc-plugin` tests.
Refs #1.

I set up a VSCode configuration and roughly documented how to debug with Chrome / generic Node inspectors. The VSCode configuration is simpler to use, but breakpoints seem to trigger unreliably. I believe the sourcemaps are working correctly as VSCode has no trouble resolving back to the original source. But sometimes, breakpoints will hit on one attempt and then not on the next when I don't think I've changed anything.