Cleans generated files during build.
Refs #1.

This deletes generated files prior to building them. Without this, it is possible that a file can be left over if a new version is not generated. This can happen for example, if a source file is deleted, then a previous build's output will still be in the output directory and possibly conflict with the build.

This should make things more hermetic and make it less likely that previous bad builds will affect future builds.