Includes name as type parameter to `MethodDescriptor`.
Refs #1.

This allows type inference to be a little stronger by having a string literal type as the name of the method, rather than the entire set of all strings.

With this change, generated `.d.ts` files with service descriptors can state the name of the method directly in the generic (as opposed to using a giant object literal and specifying all the properties manually).