Generates JavaScript and TypeScript protos descriptors.
Refs #1.

This creates a new `generators/descriptors.ts` file, which will generate JavaScript and TypeScript definitions for the descriptors for the given proto file. These descriptors contain the relevant metadata for `bxpb-runtime`, but do not implement any meaningful logic.

This was implemented as emitting two files, a `.js` file and a `.d.ts` file, rather than a single `.ts` file. The reason for this is because a single `.ts` file would need to be compiled by the consumer of the descriptors. For TypeScript projects this is trivial, but their compiler options may not be compatible with the generated code. For JavaScript projects, this requires introducing `tsc` to their build pipeline, which would be a massive and unnecessary pain. Instead, the files are emitted in a usable state already, rather than requiring a `tsc` execution.

Ideally, we would just generate a single `.ts` file internally, and then invoke `tsc` automatically with the correct options to compile it to a `.js` file and a `.d.ts` file. This could be done by adding `typescript` as a dependency and generating the code using their APIs. I opted to just write string literals as a simple and "good enough" implementation for now. Sometime in the future, it would be cool to expand this to use the real TypeScript APIs and actually invoke the compiler to generate the `.js` and `.d.ts` files.

The string literals are a bit long and complicated due to the desire to keep reasonable formatting on the generated code. I did my best not to assert on them too much in tests, and to focus on the behavior around them instead of the content in them. I think the model of dummy factories works well for these tests and I kept these as a separate file from generating client or service code to keep implementation and tests small as things inevitably become complex over time.

A few utility functions were factored out as well, as these will likely become useful when generating the client and service code.