Generates descriptors in `bxpb-protoc-plugin`.
Refs #1.

This calls the new descriptor generation code from the plugin to actuall create the new files. This is pretty simple as most of the real work is done by the descriptor generator. In particular, the tests are quite straightforward because the meaningful logic is already tested in a separate file and does not need to be repeated here.