Adds UI elements to Greeter popup.
Refs #1.

This makes the example more intuitive to users who are just trying it out to understand what's going on. A user who installed the extension would see the button in the top right of the browser and clicking it opens the popup window. They are then presented with a simple form asking for their name. Once entered, it executes the RPC and prints out the response. This way the effect is more immediately understandable than printing to the console and requiring the user to open a DevTools window to see the output.

I made an actual HTML form with a submit button and a text input which autofocuses. These are all mostly a11y concerns, but they make the form a little more usable for testing. Hopefully it doesn't obfuscate the meaningful code too much.