Adds service generation.
Refs #1.

This generates the service code for a proto source file. For the most part, this is just a call through to `bxpb-runtime`'s generic `serve()` function. However, this allows users to only worry about the generated code and not need additional mental overhead to import and use the runtime as well. The format here mostly follows the way the descriptors are generated, but a little simpler because there is just less to generate for each service.