Uses the auto-generated service in Greeter example.
Refs #1.

This updates the Greeter example app to use the new generated service code instead of calling directly into `bxpb-runtime`. This makes user code simpler by not requiring an import of the runtime and not requiring users to understand how the runtime fits into the compiler, what a service descriptor is, and how to use it with the generic `serve()` function exposed by `bxpb-runtime`.