Adds additional JSDoc comments to generated descriptor code.
Refs #1.

This more closely matches how service and client code is generated and provides inline documentation to users of the generated code.

Ideally this would `{@link ...}` to the proto source, but I have no symbols to link to due as this code itself is the service and method descriptors. Instead I just did inline code blocks with the service and method names.