Renames `bxpb-runtime` to `@bxpb/runtime`.
Refs #1.

This updates the package name to use the new NPM scope. This improves protection against typo-squatting and clearly indicates that it is an "officially" supported package.

I opted to have the directory structure as `packages/runtime/` rather than `packages/@bxpb/runtime/` as all packages should be under `@bxpb`, so it would just be extra typing and an extra directory which serves no benefit.