Adds `@fileoverview` JSDoc comments to generated code.
Refs #1.

This is generally good for documentation, but the main purpose here is to identify `bxdescriptors{.js,.d.ts}` as private, internal-only code not suitable for direct use by consumers.

This explicitly declares that the file is subject to change and is not covered by semantic versioning. This should relieve the package from having to support a particular file format and allow it to change as needed to support APIs which are actually exported.