Re-exports runtime symbols at main `@bxpb/runtime` file.
Refs #1.

This allows users to import directly from `@bxpb/runtime` and clearly defines what the supported APIs is (everything exported at that location).

Currently all the exported symbols are only used by generated code and are considered "private". I put these under `internalOnlyDoNotDependOrElse` to be clear that these APIs are not explicitly supported and are not fit for direct use by end-users.

I also chose to clear out the README for `@bxpb/runtime` because I don't want to advertise the existence of these APIs.

I couldn't find a good way of re-exporting types under a namespace. It seems like `import export Foo` syntax is supposed to be used here, but I simply could not get that to work. Instead, the generics need to be re-declared a bit awkwardly, but it shouldn't be a real problem.