Removes generic type from `ServiceDescriptor`.
Refs #1.

This type has a generic `Implementation` parameter which itself is a subtype of `ServiceDescriptor`, making this awkwardly recursive in a very unnecessary fashion.

I believe this was originally done while following similar types in gRPC generated code, however this simply isn't necessary. An object literal is already a subtype of `Record<K, V>`, so the `methods` property of `ServiceDescriptor` was ok without needing the generic. This makes the types overall simpler and easier to work with.