Adds empty assertions to type tests.
Refs #1.

These tests only exist to cause compile-errors if their respective types are not exported. Since they are compile-time checks, there is nothing to do at runtime and they were emitting warnings that no expections were present. Adding `expect().nothing()` suppresses the warning.

Also took the liberty to use `any` as a generic type argument to make the test less brittle to changes, since all it cares about is that the type is exported.