Removes `npm ci` from `prepack`.
Refs #1.

I just realized the `prepack` hook is run from Lerna, so the only way this is executed, is if the user has already run `npm install` (or `npm ci`). As a result, deleting and resetting `node_modules/` mid-execution of Lerna is likely going to cause more problems than it will solve.